% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateDMR.R
\name{annotateDMR}
\alias{annotateDMR}
\title{annotateDMR}
\usage{
annotateDMR(input, col_names, output = NULL)
}
\arguments{
\item{input}{a character string specifying the name and path of the input file (.csv)}

\item{col_names}{a character vector specifying the column names of choromosome, starting position, ending position
Should be in this format: c(chr, start, end, strand)}

\item{output}{A character string specifying the name and path of output file
the default path is as same as the input file
the default output file name is the input file name prefixed with "annotated_"}
}
\value{
a dataframe of the output data
}
\description{
This function reads differentially-methylated regions (DMRs) and annotate with hg38 information
}
\examples{
\dontrun{
df <- annotateDMR(
input = here::here("./Results/Statistics/DMR/DMR_5hmc_2018_SI_3.csv"),
col_names = c("chr", "start", "end", "strand"),
output = NULL
)
}

}
\author{
Tsunghan Hsieh
}
