% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateDMR.R
\name{.get_annotated_gr}
\alias{.get_annotated_gr}
\title{.get_annotated_gr}
\usage{
.get_annotated_gr(df, col_names, annots)
}
\arguments{
\item{col_names}{a character vector specifying the column names of return dataframe}

\item{annots}{a character vector specifying the annotation type
Should be in this format: c("mm10_cpg_islands", "mm10_cpg_shores","mm10_cpg_shelves", "mm10_cpg_inter")}

\item{input}{a character string specifying the input dataframe}
}
\value{
a dataframe of annotated file
}
\description{
This function annotate differentially-methylated regions (DMRs) with given genome information
}
\examples{
\dontrun{
df <- .get_annotated_gr(
df = input_df_pos,
col_names = col_names
)
}
}
\author{
Tsunghan Hsieh
}
