% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFittedResultByOneFactor.R
\name{getFittedResultByOneFactor}
\alias{getFittedResultByOneFactor}
\title{getFittedResultByOneFactor}
\usage{
getFittedResultByOneFactor(count, sample_sheet, sample_col, condition_col)
}
\arguments{
\item{count}{a character specifying the input count matrix,
the first column should be y and the second column should be x}

\item{sample_sheet}{a character specifying the input sample sheet,
column names of sample_sheet must contain sample_col and conditon_col}

\item{sample_col}{a character specifying the column name in sample_sheet representing the sample name}

\item{condition_col}{a character specifying the column name in sample_sheet representing the testing conditions}
}
\value{
return a dataframe speciyfing the beta and adjusted p value
}
\description{
This function fits data to the following linear model: log10(y+1) ~ x
And return the slope data and p value for the slope
As it is a multiple test, the p value will be adjusted by BH method
}
\examples{
\dontrun{
df <- getFittedResultByOneFactor(
count = count,
sample_sheet = sampleSheet_subset,
sample_col = "sample",
condition_col = "Weeks")
}
}
\author{
Tsunghan Hsieh
}
