% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savePheatPdf.R
\name{savePheatmapPdf}
\alias{savePheatmapPdf}
\title{savePheatmapPdf}
\usage{
savePheatmapPdf(x, filename, width = 7, height = 7)
}
\description{
A R function to save pheat object into pdf
This is copied from from Stackflow:
https://stackoverflow.com/questions/43051525/how-to-draw-pheatmap-plot-to-screen-and-also-save-to-file
}
