% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateDMR.R
\name{.get_annotated_gr}
\alias{.get_annotated_gr}
\title{.get_annotated_gr}
\usage{
.get_annotated_gr(df, col_names)
}
\arguments{
\item{col_names}{a character vector specifying the column names of return dataframe}

\item{input}{a character string specifying the input dataframe}
}
\value{
a dataframe of annotated file
}
\description{
This function annotate differentially-interacted regions (DIRs) with given genome information
}
\examples{
\dontrun{
df <- .get_annotated_gr(
df = input_df_pos,
col_names = col_names
)
}
}
\author{
Tsunghan Hsieh
}
